Moved Java Advanced Exercises to Correction_Java.java
